LayoutTests/editing/selection/ios/select-content-from-different-flow-002.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
<meta charset="utf-8">
<head>
<script src="../../../resources/ui-helper.js"></script>
<script src="../../../resources/js-test.js"></script>
<style>
    body {
        margin: 0;
    }

    #absolute {
        width: 100vw;
        height: 50px;
        position: absolute;
        left: 0;
        top: -100px;
    }

    #container {
        position: relative;
        margin-top: 150px;
    }
</style>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("When selecting content from different flows, i.e absolute or static, the highlight rects should not be combined if they are not adjacent.");

    var container = document.getElementById("container");
    await UIHelper.longPressElement(container);
    await UIHelper.waitForSelectionToAppear();
    getSelection().selectAllChildren(container);
    await UIHelper.waitForSelectionToAppear();
    selectionRects = await UIHelper.getUISelectionViewRects();

    shouldBe("selectionRects.length", "2");
    staticLineTop = selectionRects[0].top;
    absoluteLineBottom = selectionRects[1].top + selectionRects[0].height;
    shouldBeGreaterThan("staticLineTop", "absoluteLineBottom");

    finishJSTest();
});
</script>
</head>
<body>
    <div id="container">
        <div id="content">static</div>
        <div id="absolute">absolute</div>
    </div>
</body>
</html>
